refactor: web3 state async status
